pyproject.toml
[tool.poetry]
name = "tdd_fastapi"
version = "0.1.0"
description = ""
authors = ["Piotr Sierkin <psierkin@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "0.70"
uvicorn = "0.15"
asyncpg = "0.25"
tortoise-orm = "0.17.8"
fastapi-crudrouter = "^0.8.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
